% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_checks.R
\name{check_cells}
\alias{check_cells}
\title{Check specified cells against object}
\usage{
check_cells(cell_names, seurat_object)
}
\arguments{
\item{cell_names}{Character vector of cell names}

\item{seurat_object}{An object of type 'Seurat'}
}
\value{
No return- function will stop if supplied cell list is not in object.
}
\description{
Check specified cells against object
}
