% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_checks.R
\name{check_metadata}
\alias{check_metadata}
\title{Check metadata variables against Seurat object}
\usage{
check_metadata(seurat_object, varnames)
}
\arguments{
\item{seurat_object}{An object of type 'Seurat'}

\item{varnames}{Character vector of metadata names}
}
\value{
No return- function will stop if supplied metadata list is not in object.
}
\description{
Check metadata variables against Seurat object
}
