% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gwc_parse_args}
\alias{gwc_parse_args}
\title{Wrap optparse's parse_args() to add support for mandatory arguments}
\usage{
gwc_parse_args(option_list, mandatory = c())
}
\arguments{
\item{option_list}{List of OptionParserOption objects}

\item{mandatory}{Character vector of mandatory parameters}
}
\value{
a list containing option values, as per parse_args()
}
\description{
Wrap optparse's parse_args() to add support for mandatory arguments
}
