% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{parse_numeric}
\alias{parse_numeric}
\title{Parse numeric parameters and make sure vectors are the right length}
\usage{
parse_numeric(opt, varname, val_for_na = NA, length = 1)
}
\arguments{
\item{opt}{A list containing option values, as per parse_args()}

\item{varname}{Variable name to check}

\item{val_for_na}{Value to replicate to specified length in the case of NA}

\item{length}{The length of vector we should obtain}
}
\value{
Numeric vector
}
\description{
Sometimes a comma-separated string field needs to be parsed to a numeric vector.
}
