% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_processing.R
\name{split_string}
\alias{split_string}
\title{Split a string to a character vector}
\usage{
split_string(x, sep = ",")
}
\arguments{
\item{x}{String to be split}

\item{sep}{Separator character}
}
\value{
Character vector
}
\description{
A trivial wrapper around strsplit() so it returns a clean character vector
}
\examples{
> split_string('one,two')
"one" "two"
}
